Update App.js
Added feature that automatically adds Planned Start and End Dates at feature level. The Planned StartDate is calculated as the start date of the earliest iteration of child leaf node stories and the Planned EndDate is calculated as the End Date of the latest iteration from the child leaf node stories of the parent.